- makebook2015をplug-inのような処理に変更。 - BookMoves::size() 		// 指し		// 指し手の数は、MAX_MOVESであるので、intの範囲で十分であるのだが、 		// std::vector.size()とコンパチでないと色んなところで警告が出てうざいのでsize_tにしておく。 - gensfen2019.cpp削除。learn.cppに含める。 	- learnコマンド、コマンドごとにファイルを分けるなり整理したいが、動作検証が大変なので学習回すときにやる。 - book/makebook.cpp 追加。 - NNUEでClangでのコンパイル時に警告でてたの修正。 - 解説.txt 内容が古かったので修正。

+++ source/book/book.cpp
@@ -942,7 +942,7 @@ namespace Book
 		{
 			int pv_moves = (int)Options["BookPvMoves"];
 
-			for (int i = 0; i < move_list.size() ; ++ i)
+			for (size_t i = 0; i < move_list.size() ; ++ i)
 			{
 				// PVを構築する。pv_movesで指定された手数分だけ表示する。
 				// bestMoveを指した局面でさらに定跡のprobeを行なって…。
